move trait impls to a separate module; reorder imports with rustfmt
